// src/lib/api/model/createLeadDtoStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */

export type CreateLeadDtoStatus =
  (typeof CreateLeadDtoStatus)[keyof typeof CreateLeadDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateLeadDtoStatus = {
  NEW: "NEW",
  QUALIFYING: "QUALIFYING",
  READY_TO_PAY: "READY_TO_PAY",
  HANDED_OFF: "HANDED_OFF",
  UNQUALIFIED: "UNQUALIFIED",
  WON: "WON",
  LOST: "LOST",
} as const;